+++ textblob_de/__init__.py
@@ -13,7 +13,7 @@ from textblob_de.tokenizers import NLTKPunktTokenizer, PatternTokenizer
 from textblob_de.parsers import PatternParser
 from textblob_de.sentiments import PatternAnalyzer
 
-__version__ = '0.4.2a1'
+__version__ = '0.4.2a2'
 
 __author__ = 'Markus Killer'
 __license__ = "MIT"