+++ textblob_de/taggers.py
@@ -43,11 +43,19 @@ class PatternTagger(BaseTagger):
         :class:`PatternTokenizer() <textblob_de.tokenizers.PatternTokenizer>`.
     :param include_punc: (optional) whether to include punctuation as separate tokens.
         Default to ``False``.
+    :param encoding: (optional) Input string encoding. (Default ``utf-8``)
+    :param tagset: (optional) Penn Treebank II (default) or ('penn'|'universal'|'stts').
     '''
 
-    def __init__(self, tokenizer=None, include_punc=False):
+    def __init__(self, 
+                 tokenizer=None, 
+                 include_punc=False,
+                 encoding='utf-8',
+                 tagset=None):
         self.tokenizer = tokenizer if tokenizer is not None else PatternTokenizer()
         self.include_punc = include_punc
+        self.encoding = encoding
+        self.tagset = tagset if tagset is not None else None
 
     def tag(self, sentence, tokenize=True):
         """Tag a string `sentence`.
@@ -76,7 +84,9 @@ class PatternTagger(BaseTagger):
             sentence = _tokenized
         # Sentence is tokenized before it is passed on to pattern.de.tag
         # (i.e. it is either submitted tokenized or if )
-        _tagged = pattern_tag(sentence, tokenize=False)
+        _tagged = pattern_tag(sentence, tokenize=False,
+                              encoding=self.encoding, 
+                              tagset=self.tagset)
         if self.include_punc:
             return _tagged
         else: