+++ HISTORY.rst
@@ -7,6 +7,7 @@ Changelog
 * Removed dependency on `NLTK, <https://github.com/nltk/nltk/>`_ as it already is a `TextBlob <http://textblob.readthedocs.org/en/dev/>`_ dependency
 * (update 13/01/2015) `NLTK Issue #824 <https://github.com/nltk/nltk/issues/824>`_ fixed, workaround removed
 * Temporary workaround for `NLTK Issue #824 <https://github.com/nltk/nltk/issues/824>`_ for tox/Travis-CI
+* Enable ``pattern`` tagset conversion (``'penn'|'universal'|'stts'``) for ``PatternTagger``.
 
 0.4.1 (03/10/2014)
 ++++++++++++++++++